Finalizando app - Ignite teams

+++ igniteTeams/src/storage/group/groupCreate.ts
@@ -3,11 +3,20 @@ import AsyncStorage from '@react-native-async-storage/async-storage'
 import { GROUP_COLECTION } from '../storageConfig'
 import { groupsGetAll } from './groupsGetAll'
 
+import { AppError } from '@utils/AppError'
+
 export async function groupCreate(newGroupName:string){
 
     try{
-        const storagedGroups = await groupsGetAll();
-        const storage = JSON.stringify([...storagedGroups ,newGroupName])
+        const storedGroups = await groupsGetAll();
+
+        const groupAlreadyExist = storedGroups.includes(newGroupName)
+
+        if(groupAlreadyExist === true){
+            throw new AppError('Já existe uma turma cadastrada com esse nome !');
+        }
+
+        const storage = JSON.stringify([...storedGroups ,newGroupName])
 
         await AsyncStorage.setItem(GROUP_COLECTION , storage)
     }
